+++ src/Tiny_Websockets_Generic/network/Teensy_Ethernet_W5x00/Teensy_Ethernet_W5x00_tcp.hpp
@@ -9,7 +9,7 @@
   
   Built by Khoi Hoang https://github.com/khoih-prog/Websockets2_Generic
   Licensed under MIT license
-  Version: 1.9.1
+  Version: 1.10.0
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -36,7 +36,8 @@
   1.8.0   K Hoang      03/10/2021 Add support to RP2040, using WiFiNINA, such as Nano_RP2040_Connect
   1.8.1   K Hoang      12/10/2021 Update `platform.ini` and `library.json`
   1.9.0   K Hoang      30/11/2021 Auto detect ESP32 core version. Fix bug in examples
-  1.9.1   K Hoang      17/12/2021 Fix QNEthernet TCP interface *****************************************************************************************************************************/
+  1.9.1   K Hoang      17/12/2021 Fix QNEthernet TCP interface
+  1.10.0  K Hoang      18/12/2021 Supporting case-insensitive headers, according to RFC2616 *****************************************************************************************************************************/
  
 #pragma once