+++ src/WebSockets2_Generic.h
@@ -9,7 +9,7 @@
   
   Built by Khoi Hoang https://github.com/khoih-prog/Websockets2_Generic
   Licensed under MIT license
-  Version: 1.9.1
+  Version: 1.10.0
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -37,6 +37,7 @@
   1.8.1   K Hoang      12/10/2021 Update `platform.ini` and `library.json`
   1.9.0   K Hoang      30/11/2021 Auto detect ESP32 core version. Fix bug in examples
   1.9.1   K Hoang      17/12/2021 Fix QNEthernet TCP interface
+  1.10.0  K Hoang      18/12/2021 Supporting case-insensitive headers, according to RFC2616
  *****************************************************************************************************************************/
 
 #pragma once
@@ -44,13 +45,15 @@
 #ifndef _WEBSOCKETS2_GENERIC_H
 #define _WEBSOCKETS2_GENERIC_H
 
-#define WEBSOCKETS2_GENERIC_VERSION     "WebSockets2_Generic v1.9.1"
+#define WEBSOCKETS2_GENERIC_VERSION     "WebSockets2_Generic v1.10.0"
 
 #define WEBSOCKETS2_GENERIC_VERSION_MAJOR    1
-#define WEBSOCKETS2_GENERIC_VERSION_MINOR    9
-#define WEBSOCKETS2_GENERIC_VERSION_PATCH    1
+#define WEBSOCKETS2_GENERIC_VERSION_MINOR    10
+#define WEBSOCKETS2_GENERIC_VERSION_PATCH    0
 
-#define WEBSOCKETS2_GENERIC_VERSION_INT      1009001
+#define WEBSOCKETS2_GENERIC_VERSION_INT      1010000
+
+/////////////////////////////////////////////////////
 
 #include "Tiny_Websockets_Generic/message.hpp"
 #include "Tiny_Websockets_Generic/client.hpp"