+++ src/WebSockets2_Generic_Client.hpp
@@ -9,7 +9,7 @@
   
   Built by Khoi Hoang https://github.com/khoih-prog/Websockets2_Generic
   Licensed under MIT license
-  Version: 1.9.1
+  Version: 1.10.0
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -37,6 +37,7 @@
   1.8.1   K Hoang      12/10/2021 Update `platform.ini` and `library.json`
   1.9.0   K Hoang      30/11/2021 Auto detect ESP32 core version. Fix bug in examples
   1.9.1   K Hoang      17/12/2021 Fix QNEthernet TCP interface
+  1.10.0  K Hoang      18/12/2021 Supporting case-insensitive headers, according to RFC2616
  *****************************************************************************************************************************/
 
 #ifndef _WEBSOCKETS2_GENERIC_CLIENT_H
@@ -62,6 +63,8 @@ namespace websockets2_generic
     // Empty
   }
   
+  /////////////////////////////////////////////////////////
+  
   WebsocketsClient::WebsocketsClient(std::shared_ptr<network2_generic::TcpClient> client) :
     _client(client),
     _endpoint(client),
@@ -72,6 +75,8 @@ namespace websockets2_generic
   {
     // Empty
   }
+  
+  /////////////////////////////////////////////////////////
   
   WebsocketsClient::WebsocketsClient(const WebsocketsClient& other) :
     _client(other._client),
@@ -87,6 +92,8 @@ namespace websockets2_generic
     const_cast<WebsocketsClient&>(other)._connectionOpen = false;
   }
   
+  /////////////////////////////////////////////////////////
+  
   WebsocketsClient::WebsocketsClient(const WebsocketsClient&& other) :
     _client(other._client),
     _endpoint(other._endpoint),
@@ -100,6 +107,8 @@ namespace websockets2_generic
     const_cast<WebsocketsClient&>(other)._client = nullptr;
     const_cast<WebsocketsClient&>(other)._connectionOpen = false;
   }
+  
+  /////////////////////////////////////////////////////////
   
   WebsocketsClient& WebsocketsClient::operator=(const WebsocketsClient& other)
   {
@@ -116,8 +125,11 @@ namespace websockets2_generic
     // delete other's client
     const_cast<WebsocketsClient&>(other)._client = nullptr;
     const_cast<WebsocketsClient&>(other)._connectionOpen = false;
+    
     return *this;
   }
+  
+  /////////////////////////////////////////////////////////
   
   WebsocketsClient& WebsocketsClient::operator=(const WebsocketsClient&& other)
   {
@@ -134,8 +146,11 @@ namespace websockets2_generic
     // delete other's client
     const_cast<WebsocketsClient&>(other)._client = nullptr;
     const_cast<WebsocketsClient&>(other)._connectionOpen = false;
+    
     return *this;
   }
+  
+  /////////////////////////////////////////////////////////
   
   // KH
   
@@ -150,18 +165,22 @@ namespace websockets2_generic
     }
   }
   
+  /////////////////////////////////////////////////////////
+  
   WSString WebsocketsClient::getAuthorization(void)
   {
     return base64Authorization;
   }
   
-  //////
+  /////////////////////////////////////////////////////////
   
   struct HandshakeRequestResult
   {
     WSString requestStr;
     WSString expectedAcceptKey;
   };
+  
+  /////////////////////////////////////////////////////////
   
   bool shouldAddDefaultHeader(const std::string& keyWord, const std::vector<std::pair<WSString, WSString>>& customHeaders)
   {
@@ -176,53 +195,56 @@ namespace websockets2_generic
     return true;
   }
   
-  HandshakeRequestResult generateHandshake(const WSString& host, const WSString& uri,
-      const std::vector<std::pair<WSString, WSString>>& customHeaders) {
+  /////////////////////////////////////////////////////////
   
+  // KH 
+  HandshakeRequestResult generateHandshake(const WSString& host, const WSString& uri,
+      const std::vector<std::pair<WSString, WSString>>& customHeaders) 
+  {
     WSString key = crypto2_generic::base64Encode(crypto2_generic::randomBytes(16));
   
     WSString handshake = "GET " + uri + " HTTP/1.1\r\n";
-    handshake += "Host: " + host + "\r\n";
-    handshake += "Sec-WebSocket-Key: " + key + "\r\n";
+    handshake += HEADER_HOST + host + HEADER_HOST_RN;
+    handshake += HEADER_WS_KEY + key + HEADER_HOST_RN;
   
     for (const auto& header : customHeaders)
     {
-      handshake += header.first + ": " + header.second + "\r\n";
+      handshake += header.first + ": " + header.second + HEADER_HOST_RN;
     }
   
-    if (shouldAddDefaultHeader("Upgrade", customHeaders))
+    if (shouldAddDefaultHeader(HEADER_UPGRADE_NORMAL, customHeaders))
     {
-      handshake += "Upgrade: websocket\r\n";
+      handshake += HEADER_UPGRADE_WS;
     }
   
-    if (shouldAddDefaultHeader("Connection", customHeaders))
+    if (shouldAddDefaultHeader(HEADER_CONNECTION_NORMAL, customHeaders))
     {
-      handshake += "Connection: Upgrade\r\n";
+      handshake += HEADER_CONNECTION_UPGRADE;
     }
   
-    if (shouldAddDefaultHeader("Sec-WebSocket-Version", customHeaders))
+    if (shouldAddDefaultHeader(HEADER_WS_VERSION_NORMAL, customHeaders))
     {
-      handshake += "Sec-WebSocket-Version: 13\r\n";
+      handshake += HEADER_WS_VERSION_13;
     }
   
-    if (shouldAddDefaultHeader("User-Agent", customHeaders))
+    if (shouldAddDefaultHeader(HEADER_USER_AGENT_NORMAL, customHeaders))
     {
-      handshake += "User-Agent: TinyWebsockets Client\r\n";
+      handshake += HEADER_USER_AGENT_VALUE;
     }
   
-    if (shouldAddDefaultHeader("Origin", customHeaders))
+    if (shouldAddDefaultHeader(HEADER_ORIGIN_NORMAL, customHeaders))
     {
-      handshake += "Origin: https://github.com/khoih-prog/Websockets2_Generic\r\n";
+      handshake += HEADER_ORIGIN_VALUE;
     }
   
-    handshake += "\r\n";
+    handshake += HEADER_HOST_RN;
   
     HandshakeRequestResult result;
     result.requestStr = handshake;
     
 
     // KH
-    LOGWARN1("WebsocketsClient::generateHandshake: handshake =", internals2_generic::fromInternalString(handshake));
+    LOGINFO1("WebsocketsClient::generateHandshake: handshake =", internals2_generic::fromInternalString(handshake));
     ////// 
 
   
@@ -234,66 +256,66 @@ namespace websockets2_generic
     return result;
   }
   
-  // KH
+  /////////////////////////////////////////////////////////
   
   HandshakeRequestResult generateHandshake(const WSString& host, const WSString& uri,
-      const std::vector<std::pair<WSString, WSString>>& customHeaders, const WSString& base64Authorization) {
-  
+      const std::vector<std::pair<WSString, WSString>>& customHeaders, const WSString& base64Authorization) 
+  {  
     WSString key = crypto2_generic::base64Encode(crypto2_generic::randomBytes(16));
   
     WSString handshake = "GET " + uri + " HTTP/1.1\r\n";
-    handshake += "Host: " + host + "\r\n";
-    handshake += "Sec-WebSocket-Key: " + key + "\r\n";
+    handshake += HEADER_HOST + host + HEADER_HOST_RN;
+    handshake += HEADER_WS_KEY + key + HEADER_HOST_RN;
   
     for (const auto& header : customHeaders)
     {
-      handshake += header.first + ": " + header.second + "\r\n";
+      handshake += header.first + ": " + header.second + HEADER_HOST_RN;
     }
   
-    if (shouldAddDefaultHeader("Upgrade", customHeaders))
+    if (shouldAddDefaultHeader(HEADER_UPGRADE_NORMAL, customHeaders))
     {
-      handshake += "Upgrade: websocket\r\n";
+      handshake += HEADER_UPGRADE_WS;
     }
   
-    if (shouldAddDefaultHeader("Connection", customHeaders))
+    if (shouldAddDefaultHeader(HEADER_CONNECTION_NORMAL, customHeaders))
     {
-      handshake += "Connection: Upgrade\r\n";
+      handshake += HEADER_CONNECTION_UPGRADE;
     }
   
-    if (shouldAddDefaultHeader("Sec-WebSocket-Version", customHeaders))
+    if (shouldAddDefaultHeader(HEADER_WS_VERSION_NORMAL, customHeaders))
     {
-      handshake += "Sec-WebSocket-Version: 13\r\n";
+      handshake += HEADER_WS_VERSION_13;
     }
   
-    if (shouldAddDefaultHeader("User-Agent", customHeaders))
+    if (shouldAddDefaultHeader(HEADER_USER_AGENT_NORMAL, customHeaders))
     {
-      handshake += "User-Agent: TinyWebsockets Client\r\n";
+      handshake += HEADER_USER_AGENT_VALUE;
     }
   
     // KH
     //if (WebsocketsClient::base64Authorization.size() > 0)
     {
-      handshake += "Authorization: Basic ";
-      handshake += base64Authorization + "\r\n";
+      handshake += HEADER_AUTH_BASIC;  
+      handshake += base64Authorization + HEADER_HOST_RN;
       
       // KH
-      LOGWARN1("WebsocketsClient::generateHandshake: base64Authorization =", internals2_generic::fromInternalString(base64Authorization));           
+      LOGINFO1("WebsocketsClient::generateHandshake: base64Authorization =", internals2_generic::fromInternalString(base64Authorization));
       //////
     }
 
-    if (shouldAddDefaultHeader("Origin", customHeaders))
+    if (shouldAddDefaultHeader(HEADER_ORIGIN_NORMAL, customHeaders))
     {
-      handshake += "Origin: https://github.com/khoih-prog/Websockets2_Generic\r\n";
+      handshake += HEADER_ORIGIN_VALUE;
     }
     //////
   
-    handshake += "\r\n";
+    handshake += HEADER_HOST_RN;
   
     HandshakeRequestResult result;
     result.requestStr = handshake;
     
     // KH
-    LOGWARN1("WebsocketsClient::generateHandshake: handshake =", internals2_generic::fromInternalString(handshake));
+    LOGINFO1("WebsocketsClient::generateHandshake: handshake =", internals2_generic::fromInternalString(handshake));
     ////// 
   
   #ifndef _WS_CONFIG_SKIP_HANDSHAKE_ACCEPT_VALIDATION
@@ -303,19 +325,24 @@ namespace websockets2_generic
     // KH, Don't need return std::move(result);
     return result;
   }
-  //////
   
+  /////////////////////////////////////////////////////////
+
   bool isWhitespace(char ch)
   {
     return ch == ' ' || ch == '\t' || ch == '\r' || ch == '\n';
   }
   
+  /////////////////////////////////////////////////////////
+
   struct HandshakeResponseResult
   {
     bool isSuccess;
     WSString serverAccept;
   };
   
+  /////////////////////////////////////////////////////////
+
   bool isCaseInsensetiveEqual(const WSString lhs, const WSString rhs)
   {
     if (lhs.size() != rhs.size())
@@ -333,6 +360,8 @@ namespace websockets2_generic
     return true;
   }
   
+  /////////////////////////////////////////////////////////
+
   HandshakeResponseResult parseHandshakeResponse(std::vector<WSString> responseHeaders)
   {
     bool didUpgradeToWebsockets = false, isConnectionUpgraded = false;
@@ -346,19 +375,19 @@ namespace websockets2_generic
       WSString value = header.substr(colonIndex + 2); // +2 (ignore space and ':')
       
       // KH
-      LOGWARN1("WebsocketsClient::generateHandshake: key =", internals2_generic::fromInternalString(key));
-      LOGWARN1("WebsocketsClient::generateHandshake: value =", internals2_generic::fromInternalString(value));
+      LOGDEBUG1("WebsocketsClient::generateHandshake: key =", internals2_generic::fromInternalString(key));
+      LOGDEBUG1("WebsocketsClient::generateHandshake: value =", internals2_generic::fromInternalString(value));
       ////// 
   
-      if (isCaseInsensetiveEqual(key, "upgrade"))
+      if (isCaseInsensetiveEqual(key, HEADER_UPGRADE_LOWER_CASE))
       {
-        didUpgradeToWebsockets = isCaseInsensetiveEqual(value, "Websocket");
+        didUpgradeToWebsockets = isCaseInsensetiveEqual(value, HEADER_WEBSOCKET_LOWER_CASE);
       }
-      else if (isCaseInsensetiveEqual(key, "connection"))
+      else if (isCaseInsensetiveEqual(key, HEADER_CONNECTION_LOWER_CASE))
       {
-        isConnectionUpgraded = isCaseInsensetiveEqual(value, "upgrade");
+        isConnectionUpgraded = isCaseInsensetiveEqual(value, HEADER_UPGRADE_LOWER_CASE);
       }
-      else if (isCaseInsensetiveEqual(key, "Sec-WebSocket-Accept"))
+      else if (isCaseInsensetiveEqual(key, WS_ACCEPT_NORMAL))
       {
         serverAccept = value;
       }
@@ -371,6 +400,8 @@ namespace websockets2_generic
     return result;
   }
   
+  /////////////////////////////////////////////////////////
+
   bool doestStartsWith(WSString str, WSString prefix)
   {
     if (str.size() < prefix.size())
@@ -390,13 +421,15 @@ namespace websockets2_generic
     }
     
     // KH
-    LOGINFO1("WebsocketsClient::doestStartsWith: str =", internals2_generic::fromInternalString(str));
-    LOGINFO1("WebsocketsClient::doestStartsWith: prefix =", internals2_generic::fromInternalString(prefix));
+    LOGDEBUG1("WebsocketsClient::doestStartsWith: str =", internals2_generic::fromInternalString(str));
+    LOGDEBUG1("WebsocketsClient::doestStartsWith: prefix =", internals2_generic::fromInternalString(prefix));
     ////// 
           
     return true;
   }
   
+  /////////////////////////////////////////////////////////
+
   // KH, New in v1.0.5 (sync with v0.4.18)
   void WebsocketsClient::upgradeToSecuredConnection()
   {
@@ -475,13 +508,16 @@ namespace websockets2_generic
     //////
   #endif //_WS_CONFIG_NO_SSL
   }
-  //////
+  
+  /////////////////////////////////////////////////////////
   
   void WebsocketsClient::addHeader(const WSInterfaceString key, const WSInterfaceString value)
   {
     _customHeaders.push_back({internals2_generic::fromInterfaceString(key), internals2_generic::fromInterfaceString(value)});
   }
   
+  /////////////////////////////////////////////////////////
+
   bool WebsocketsClient::connect(WSInterfaceString _url)
   {
     WSString url = internals2_generic::fromInterfaceString(_url);
@@ -568,6 +604,8 @@ namespace websockets2_generic
            );
   }
   
+  /////////////////////////////////////////////////////////
+
   bool WebsocketsClient::connect(WSInterfaceString host, int port, WSInterfaceString path)
   {
     // KH
@@ -591,7 +629,7 @@ namespace websockets2_generic
     auto handshake = generateHandshake(internals2_generic::fromInterfaceString(host), 
                         internals2_generic::fromInterfaceString(path), _customHeaders, base64Authorization);
                         
-    LOGINFO1("WebsocketsClient::connect: base64Authorization =", internals2_generic::fromInternalString(base64Authorization));             
+    LOGINFO1("WebsocketsClient::connect: base64Authorization =", internals2_generic::fromInternalString(base64Authorization));
     //////
     
     // KH
@@ -647,7 +685,7 @@ namespace websockets2_generic
         return false;
       }
   
-      if (line == "\r\n")
+      if (line == HEADER_HOST_RN)
         break;
   
       // remove /r/n from line end
@@ -685,11 +723,24 @@ namespace websockets2_generic
     return true;
   }
   
+  /////////////////////////////////////////////////////////
+  
+  bool WebsocketsClient::connectSecure(WSInterfaceString host, int port, WSInterfaceString path) 
+  {
+    upgradeToSecuredConnection();
+
+    return connect(host, port, path);
+  }
+  
+  /////////////////////////////////////////////////////////
+
   void WebsocketsClient::onMessage(MessageCallback callback)
   {
     this->_messagesCallback = callback;
   }
   
+  /////////////////////////////////////////////////////////
+
   void WebsocketsClient::onMessage(PartialMessageCallback callback)
   {
     this->_messagesCallback = [callback](WebsocketsClient&, WebsocketsMessage msg)
@@ -698,11 +749,15 @@ namespace websockets2_generic
     };
   }
   
+  /////////////////////////////////////////////////////////
+
   void WebsocketsClient::onEvent(EventCallback callback)
   {
     this->_eventsCallback = callback;
   }
   
+  /////////////////////////////////////////////////////////
+
   void WebsocketsClient::onEvent(PartialEventCallback callback)
   {
     this->_eventsCallback = [callback](WebsocketsClient&, WebsocketsEvent event, WSInterfaceString data)
@@ -711,6 +766,8 @@ namespace websockets2_generic
     };
   }
   
+  /////////////////////////////////////////////////////////
+
   bool WebsocketsClient::poll()
   {
     bool messageReceived = false;
@@ -762,7 +819,8 @@ namespace websockets2_generic
     else
       return {};  
   }
-  //////
+  
+  /////////////////////////////////////////////////////////
   
   WebsocketsMessage WebsocketsClient::readBlocking()
   {
@@ -780,23 +838,31 @@ namespace websockets2_generic
     return {};
   }
   
+  /////////////////////////////////////////////////////////
+
   bool WebsocketsClient::send(const WSInterfaceString& data)
   {
     auto str = internals2_generic::fromInterfaceString(data);
     return this->send(str.c_str(), str.size());
   }
   
+  /////////////////////////////////////////////////////////
+
   bool WebsocketsClient::send(const WSInterfaceString&& data)
   {
     auto str = internals2_generic::fromInterfaceString(data);
     return this->send(str.c_str(), str.size());
   }
   
+  /////////////////////////////////////////////////////////
+
   bool WebsocketsClient::send(const char* data)
   {
     return this->send(data, strlen(data));
   }
   
+  /////////////////////////////////////////////////////////
+
   bool WebsocketsClient::send(const char* data, const size_t len)
   {
     if (available())
@@ -824,15 +890,20 @@ namespace websockets2_generic
                );
       }
     }
+    
     return false;
   }
   
+  /////////////////////////////////////////////////////////
+
   bool WebsocketsClient::sendBinary(WSInterfaceString data)
   {
     auto str = internals2_generic::fromInterfaceString(data);
     return this->sendBinary(str.c_str(), str.size());
   }
   
+  /////////////////////////////////////////////////////////
+
   bool WebsocketsClient::sendBinary(const char* data, const size_t len)
   {
     if (available())
@@ -863,6 +934,8 @@ namespace websockets2_generic
     return false;
   }
   
+  /////////////////////////////////////////////////////////
+
   bool WebsocketsClient::stream(const WSInterfaceString data)
   {
     if (available() && this->_sendMode == SendMode_Normal)
@@ -877,7 +950,8 @@ namespace websockets2_generic
     return false;
   }
   
-  
+  /////////////////////////////////////////////////////////
+
   bool WebsocketsClient::streamBinary(const WSInterfaceString data)
   {
     if (available() && this->_sendMode == SendMode_Normal)
@@ -892,6 +966,8 @@ namespace websockets2_generic
     return false;
   }
   
+  /////////////////////////////////////////////////////////
+
   bool WebsocketsClient::end(const WSInterfaceString data)
   {
     if (available() && this->_sendMode == SendMode_Streaming)
@@ -906,14 +982,19 @@ namespace websockets2_generic
     return false;
   }
   
+  /////////////////////////////////////////////////////////
+
   void WebsocketsClient::setFragmentsPolicy(const FragmentsPolicy newPolicy)
   {
     _endpoint.setFragmentsPolicy(newPolicy);
   }
   
+  /////////////////////////////////////////////////////////
+
   bool WebsocketsClient::available(const bool activeTest)
   {
-    if (activeTest)  {
+    if (activeTest)  
+    {
       _endpoint.ping("");
     }
   
@@ -926,9 +1007,12 @@ namespace websockets2_generic
     }
   
     this->_connectionOpen = updatedConnectionOpen;
+    
     return this->_connectionOpen;
   }
   
+  /////////////////////////////////////////////////////////
+
   bool WebsocketsClient::ping(const WSInterfaceString data)
   {
     if (available())
@@ -939,6 +1023,8 @@ namespace websockets2_generic
     return false;
   }
   
+  /////////////////////////////////////////////////////////
+
   bool WebsocketsClient::pong(const WSInterfaceString data)
   {
     if (available())
@@ -949,6 +1035,8 @@ namespace websockets2_generic
     return false;
   }
   
+  /////////////////////////////////////////////////////////
+
   void WebsocketsClient::close(const CloseReason reason)
   {
     if (available())
@@ -959,26 +1047,35 @@ namespace websockets2_generic
     }
   }
   
+  /////////////////////////////////////////////////////////
+
   CloseReason WebsocketsClient::getCloseReason() const
   {
     return _endpoint.getCloseReason();
   }
   
+  /////////////////////////////////////////////////////////
+
   void WebsocketsClient::_handlePing(const WebsocketsMessage message)
   {
     this->_eventsCallback(*this, WebsocketsEvent::GotPing, message.data());
   }
   
+  /////////////////////////////////////////////////////////
+
   void WebsocketsClient::_handlePong(const WebsocketsMessage message)
   {
     this->_eventsCallback(*this, WebsocketsEvent::GotPong, message.data());
   }
   
+  /////////////////////////////////////////////////////////
+
   void WebsocketsClient::_handleClose(const WebsocketsMessage message)
   {
     this->_eventsCallback(*this, WebsocketsEvent::ConnectionClosed, message.data());
   }
   
+  /////////////////////////////////////////////////////////
   
   #ifdef ESP8266
   void WebsocketsClient::setFingerprint(const char* fingerprint)
@@ -986,6 +1083,8 @@ namespace websockets2_generic
     this->_optional_ssl_fingerprint = fingerprint;
   }
   
+  /////////////////////////////////////////////////////////
+
   // KH, New in v1.0.5 (sync with v0.4.18)
   void WebsocketsClient::setInsecure() 
   {
@@ -997,29 +1096,38 @@ namespace websockets2_generic
     this->_optional_ssl_trust_anchors = nullptr;
     this->_optional_ssl_known_key = nullptr;
   }
+  
+  /////////////////////////////////////////////////////////
 
   void WebsocketsClient::setClientRSACert(const X509List *cert, const PrivateKey *sk) 
   {
     this->_optional_ssl_rsa_cert = cert;
     this->_optional_ssl_rsa_private_key = sk;
   }
+  
+  /////////////////////////////////////////////////////////
 
   void WebsocketsClient::setClientECCert(const X509List *cert, const PrivateKey *sk) 
   {
     this->_optional_ssl_ec_cert = cert;
     this->_optional_ssl_ec_private_key = sk;
   }
+  
+  /////////////////////////////////////////////////////////
 
   void WebsocketsClient::setTrustAnchors(const X509List *ta) 
   {
     this->_optional_ssl_trust_anchors = ta;
   }
+  
+  /////////////////////////////////////////////////////////
 
   void WebsocketsClient::setKnownKey(const PublicKey *pk) 
   {
     this->_optional_ssl_known_key = pk;
   }
-  //////
+  
+  /////////////////////////////////////////////////////////
   
   #elif defined(ESP32)
   void WebsocketsClient::setCACert(const char* ca_cert)
@@ -1027,16 +1135,22 @@ namespace websockets2_generic
     this->_optional_ssl_ca_cert = ca_cert;
   }
   
+  /////////////////////////////////////////////////////////
+
   void WebsocketsClient::setCertificate(const char* client_ca)
   {
     this->_optional_ssl_client_ca = client_ca;
   }
   
+  /////////////////////////////////////////////////////////
+
   void WebsocketsClient::setPrivateKey(const char* private_key)
   {
     this->_optional_ssl_private_key = private_key;
   }
   
+  /////////////////////////////////////////////////////////
+
   void WebsocketsClient::setInsecure()
   {
     this->_optional_ssl_ca_cert = nullptr;
@@ -1045,6 +1159,8 @@ namespace websockets2_generic
   }
   #endif
   
+  /////////////////////////////////////////////////////////
+
   WebsocketsClient::~WebsocketsClient()
   {
     if (available())
@@ -1052,6 +1168,9 @@ namespace websockets2_generic
       this->close(CloseReason_GoingAway);
     }
   }
+  
+  /////////////////////////////////////////////////////////
+
 }    //namespace websockets2_generic
 
 #endif    // _WEBSOCKETS2_GENERIC_CLIENT_H